# openapi.yaml
openapi: '3.1.0'
info:
    title: Manna AI Agent API
    version: '1.0.0'
    contact:
        name: Manna API
        url: https://github.com/Guebbit/manna
    description: |
        Local-first agentic AI platform. Exposes the agent reasoning loop, IDE-specific
        direct-LLM endpoints, file-upload helpers,
        and a multi-library PDF ingestion and semantic search system.

        **Base URL (default):** `http://localhost:3001`
servers:
    - url: http://localhost:3001
      description: Local development server
...
paths:
...
    /upload/image-classify:
        post:
            tags: [upload]
            summary: Classify or describe an uploaded image
            description: |
                Accepts a multipart file upload and classifies/describes the image using
                the model configured in `TOOL_VISION_MODEL`. Max file size: 50 MB.
            operationId: uploadImageClassify
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required: [file]
                            properties:
                                file:
                                    type: string
                                    format: binary
                                    description: 'Allowed image MIME types: image/jpeg, image/png, image/gif, image/webp.'
                                prompt:
                                    type: string
                                    description: Optional prompt to guide the classification
                                model:
                                    type: string
                                    description: Override the vision model (Ollama model name)
            responses:
                '200':
                    description: Classification result
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    result:
                                        type: string
                                        description: Model's description of the image
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'
    /upload/speech-to-text:
        post:
            tags: [upload]
            summary: Transcribe an uploaded audio file
            description: |
                Accepts a multipart audio upload and returns a transcript using
                `TOOL_STT_MODEL` (default `whisper`). Max file size: 50 MB.
            operationId: uploadSpeechToText
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required: [file]
                            properties:
                                file:
                                    type: string
                                    format: binary
                                    description: 'Allowed audio MIME types: audio/mpeg, audio/wav, audio/ogg, audio/webm, audio/mp4.'
                                model:
                                    type: string
                                    description: Override the STT model
                                language:
                                    type: string
                                    description: BCP-47 language code (e.g. "en", "fr")
                                prompt:
                                    type: string
                                    description: Optional context prompt to improve transcription accuracy
            responses:
                '200':
                    description: Transcription result
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    text:
                                        type: string
                                        description: Transcribed text
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'
    /upload/read-pdf:
        post:
            tags: [upload]
            summary: Extract text from an uploaded PDF
            description: |
                Accepts a multipart PDF upload and returns the extracted text plus page count.
                Max file size: 50 MB.
            operationId: uploadReadPdf
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required: [file]
                            properties:
                                file:
                                    type: string
                                    format: binary
                                    description: 'Allowed document MIME type: application/pdf.'
            responses:
                '200':
                    description: Extracted PDF content
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    text:
                                        type: string
                                        description: Full extracted text
                                    pages:
                                        type: integer
                                        description: Total number of pages in the PDF
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'

    # ── Info endpoints ──────────────────────────────────────────────────────────
...
components:
...
    responses:
        BadRequest:
            description: Invalid request body or parameters
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
...
        InternalError:
            description: Internal server error (LLM failure, Qdrant unavailable, etc.)
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
    schemas:
        ErrorResponse:
            type: object
            required: [error]
            properties:
                error:
                    type: string
                    description: Human-readable error message.
                    example: "Library 'scientific-american' not found"
...